backend/requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
openai
langchain-groq
duckduckgo_search==5.3.1b1
pymupdf
arxiv
wikipedia
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub
streamlit 
langchain_groq
langserve 
fastapi 
uvicorn 
langgraph 
langchain_anthropic
python-dotenv
langchain_google_community
google-cloud-discoveryengine
vertexai
chainlit 
pillow 
uvicorn
beautifulsoup4
fastapi
google-cloud-aiplatform
google-cloud-discoveryengine
langcodes
langdetect
pydantic
pyyaml
vertexai
uvicorn
langid 
google-cloud-translate
google-cloud-bigquery
db-dtypes
langchain_openai 
google-cloud-bigquery-storage 
google-cloud-vision